Cleanup: move all route code to explorer code. Make it easier to remove if need be
